Add digit separators to long numbers in numeric tooltip
Long numbers are hard to read in the numeric tooltip. To improve
readability, add language-specific digit separators to the numeric
tooltip.

Decimal numbers are grouped into chunks of three digits while
hexadecimal numbers are grouped into chunks of length four.

The digit separator is language-specific and chosen so that the number
is a valid token in the source language.