Simplify Android compiler properties
updates intermediate compiler exe input to refer to the compiler's
getInfo().exe so that paths don't need to be hardcoded in properties.

This change also makes the D8 class configurable, as local builds of the
R8 jar may point to a different class. The same class is used by
default.