Add R8 for Android
R8 is a whole-program optimizer that converts Java byte code to
optimized dex code. It can be run with the same general steps as the
existing D8 compiler, but additionally requires android.jar and the
keep-annotation library.

This change also includes a small update to dex2oat that allows for
older versions to be used in CE.

New code examples still need to be added for Android Java/Kotlin in
order to provide steps for using R8.