Improve passthrough of -march and -mcpu from compiler to llvm-mca
This patch fixes a few issues with argument passthrough that have been
discovered since #6682 was landed.

* X86 didn't work if -march=CPU was supplied.
* It wasn't possible to override flags on MCA, because the auto-detected
  flags came before user flags.
* Since compiler-explorer runs on x86 native hardware, if targetting
  aarch64, it would result in an error about the default requested
  native CPU not being available for the architecture.

X86 clang uses -march to specify the CPU, and it's invalid to supply
-march=<cpu> to LLVM-MCA, because it is interpreted as a target.
Meanwhile, LLVM-MCA requires an -mcpu argument.

Therefore, to handle x86 it's necessary to rewrite -march to -mcpu so
long as the march flag is used to supply a CPU value (i.e. not an
architecture level beginning with `x86-64`).

Per https://maskray.me/blog/2022-08-28-march-mcpu-mtune, I understand
x86 is currently the only target to have this behaviour so it needs
special casing.

Also, handle the case where --target is being passed to the compiler,
and pass this through as -mtriple.

This also makes compiler-explorer pass -mcpu=generic if an -mcpu is not
otherwise specified, with the rationale that llvm-mca running on
compiler-explorer would otherwise produce results which depend on the
physical CPU of the machine; this was undesirable because it could vary,
and the default of -mcpu=native also meant that cross-architecture would
not function by default.

Testing:

* I have been testing on an aarch64-native machine.
* I have tested targeting x86 by specifying --target and using the
  override functionality in compiler explorer to specify x86.
  I have also tested -target <arg>.
* I have been inspecting the `--debug` output to see what flags are
  passed to LLVM-MCA.
* I've checked that I can override the -mcpu in MCA by supplying
  arguments to MCA, regardless of what is passed to the compiler.
* I have checked that I can supply either -march=x86-64* or -march=<cpu>
  on the x86 target and things behave as expected.
* If -m32 is passed on x86, pass -march=x86 to llvm-mca, which is how to
  specify the 32 bit target there and can result in subtle differences.
* Also pass -march=x86 to llvm-mca if the target is i386.

Fixes: #6709