Remove '-p' from forced demangler args (#6750)
Minor added tsification

Fixes #6714 : "No demangling in opt-remarks for clang <=17"

This is probably a regression introduced by my PR #6451 - which changed
the version of the llvm-cxxfilt demanglers used. Up until llvm 18 these
demanglers did not support the `-p` switch ("Skip function parameters
and return types").
The fix here is to remove the switch (why would we want it?? We wish to
distinguish overloads in opt remarks too).