Add explicit arm64, arm32, x86, x86_64 variants of Android Java dex2oat
This was possible before in the UI by adding `--instruction-set=arm64`, but not
very discoverable, and not consistent with how we do other compilers, and
doesn't let us override the instruction set output for assembly hovertext
documentation purposes.

There are two properties in play:
- instructionSet, which Compiler Explorer uses to choose which assembly
  documentation hovertext to use. e.g. `aarch64`.
- dex2oatInstructionSet, which informs the `dex2oat --instruction-set` flag. e.g. `arm64.

These have slightly different representations for the same concept.

Towards #6818