[LoongArch] Fix llvm config for #6793 (#6840)
In pr #6793, we misconfigured the `target`. This patch addresses and
corrects the problem. Additionally, the GCC toolchain has been updated
to handle the `ld` linker’s mismatch with the latest relocation type.

<!-- THIS COMMENT IS INVISIBLE IN THE FINAL PR, BUT FEEL FREE TO REMOVE
IT
Thanks for taking the time to improve CE. We really appreciate it.
Before opening the PR, please make sure that the tests & linter pass
their checks,
  by running `make check`.
In the best case scenario, you are also adding tests to back up your
changes,
  but don't sweat it if you don't. We can discuss them at a later date.
Feel free to append your name to the CONTRIBUTORS.md file
Thanks again, we really appreciate this!
-->

Signed-off-by: Qi HU <github@spcsky.com>