[PROTOTYPE] Support chaining compiler outputs
This adds (hacky) support for chaining the output from one compiler to
the input of another. The intended use case of this is for compilers for
languages like MLIR, V, Jakt, etc that don't lower directly to machine
code but rather another IR or source language. Chaining compilers allows
for an easy way to set up a full complication pipeline (and possibly
even execute the result).

Disclaimer: I don't know much about CE internals, but this was
surprisingly easy to hack together by abusing existing events.

This prototype works by marking the compiler editor (`sourceEditorId`)
as another compiler by setting a high bit
(0x8000 - EDITOR_IS_COMPILER_OUTPUT). The compiler pane can then listen
for the `onCompileResult` event from its chained source compiler, and
use that to update its source code (and request recompilation).

This requires the compilation result to include the unfiltered output
to use as the source, along with a way of querying the output langId.

This prototype does not add a good way to get the result langId and
currently uses the monaco ID (which unfortunately does not map to CE
langIds neatly, for example, the 'mlir' monaco ID could be for 'circt'
or 'mlir'). But I think compilers could be extended to return the result
langId fairly easily.