Support kotlinc output for Android Kotlin
This change adds kotlinc as an output option for Android Kotlin, and
updates kotlin.ts to allow for input libraries by parsing library paths
and passing them in with '-cp path:path:...'.

This is intended to simplify the process of inspecting kotlinc output
when Android Kotlin is already selected so that a user doesn't have to
switch to the Kotlin compiler option. This also allows a user to view
kotlinc output with Android libraries.