Decommision the "AWS" magic compiler name
Nothing was using it, we're using other methods these
days to discover remote compilers (and have been for years),
and it was broken (see #1790).

Also cleans up some config (moving it from aws to ceProps; though
it was unused, default used only), and also some workarounds for
remote compilers that refer back to the MS compiler era (#1768).