Handle ARM's .set usage (#7322)
Treats `.set` as an "alias" for the set of labels referred to, and
doesn't filter alias-creating `.set` in used paths even if directive
filtering is on. Only works for `.set <something>, .` with an optional
`+ 0`.

Also changes a silly loop that was looping N times doing the same thing
if `inFunction && isDataDefinition`, though quite possibly that's broken
in a different way. No tests changed, anyway.

Adds a new test for the new behaviour.

Closes #1799