Mitigations for CL labels (#7320)
Some syntax highlighting improvement for CL asm lines like:

```as
int * moo DD  01H                       ; moo
```

and

```as
?mooz@@3PAFA DW 01H                                 ; mooz
```

where previously the DD (or DW DQ) wasn't syntax highlighted, making it
harder to spot the case where CL puts the first defined value of an
array on the same line as the label definition.

This isn't perfect, and might be fairly expensive, but helps mitigate
issue #1645, hopefully well enough to close it.

Tested locally with gcc and clang output, and then pasted into
Monarch/Monaco and tested with some CL output.