Arm: add missing objdumper settings and update to GNU objdump 15.1.0
Fixes #7635

Prior to this change the C++ compilers set objdumper on the Armv7/v8
category as a whole, pointing to some objdump from a gcc cross compiler.
This works because that tool expects Arm objects, and you could use the
"compile to binary object" option successfully.

For C, we only set objdumper for specific compilers. Those being the
trunk and "all features" versions. So when "compile to binary object"
was enabled it worked for trunk but never for any version before that.

The other versions were using the settings file's global objdumper
which pointed to an objdump that only expects x86 object files.

/opt/compiler-explorer/gcc-14.2.0/bin/objdump: can't disassemble for architecture UNKNOWN!
/tmp/compiler-explorer-compiler4xFvjY/output.s:     file format elf64-little

To fix this I have copied the way that the C++ settings work, into
the C settings. So all Armv7 and Armv8 clangs get their objdumper
setting from the category of compilers, rather than setting it for each compiler.

While I'm doing this, we might as well update to the objdump from
GCC 15.1.0, so that we can disassemble new instructions.

The worst that can happen is some unallocated encoding at the time
of clang X.Y is now allocated, but compilers should be using the
specific "udf" (always undefined) encodings or marking these values
as data anyway. So I don't see the harm in always using a new objdump
everywhere.