Show opt remarks in a view zone instead of as editor content (#7687)
Resolves #7662.

[View
zones](https://microsoft.github.io/monaco-editor/typedoc/interfaces/editor.IViewZone.html)
are rendered in the editor widget but don’t count as lines themselves.
Monaco doesn’t automatically grow them to fit their content (and doesn’t
extend the scroll area to fit them when the text is rendered without
line breaks), so the height must be calculated manually.

Additionally, this changes the remarks to be *above* the line they are
attached to. This is more intuitive in my opinion, but can be reverted
again if you prefer.

Comparison:

before:

![opt_view_before](https://github.com/user-attachments/assets/7cd34349-4a97-4da8-b3c4-ec197542fb35)

after:

![opt_view](https://github.com/user-attachments/assets/c731ae02-b626-44d7-a9a2-77743747733d)