Add bmpi3 (#7735)
Proposing to add the B-MPI3 library to the live site.
https://gitlab.com/correaa/boost-mpi3

Followed the steps as best as I could in based on other gitlab type
libraries (bmulti)

Feedback on improving the hooks and the organizations of files in the
repository is welcome.
The goal is to include the library by doing
#include<mpi3/communicator.hpp>.