Sort rakudo versions newest-first
by adding a (strictly speaking wrong) isSemVer=true and
setting a SemVer prop based on the year.month.patch versioning
that the releases have.

The decision to keep the "name" property instead of relying on
baseName + SemVer is to keep the naming scheme as
Rakudo $version $backend instead of having the backend in the middle.