Enumerate all logo files for webpack
When we move to Vite we won't have the `require.context`-like functionality that
webpack has. For this reason we have to enumerate all the files and import them
individually.

I will be opening a similar patch for the site template screenshot images.

The `logoUrl` and `logoUrlDark` fields have been renamed to `logoFilename` to
be more accurate for what the keys are now. Sonar file was also renamed for
consistency sake.