Add `cursor: pointer` to some clickable texts (#7863)
Before #7797, elements such as “Toggle full version output” in the
compiler version popup or the share buttons (“Short Link” etc.) would
change the cursor to a pointer on hover (like normal links do), but this
was lost when they were switched to use `link-primary`.

This PR restores the previous behaviour.