Fix #8037: avoid duplicate `compile` call (#8038)
Trying to solve the root cause:

`Editor.maybeEmitChange` calls `this.eventHub.emit('editorChange')`, and
then `Compiler.onEditorChange` calls `this.compile()` .
So the idiom
```ts
  this.maybeEmitChange();
  this.requestCompilation();
```
repeated a few times in editor.ts, causes two rapid (and redundant)
invocations of `compile`, which in turn cause the race. This PR removes
all calls to `Editor.requestCompilation` and in fact removes the
function entirely.

I'm guessing `maybeEmitChange` is a late addition to solve some other
problems, and this race was an unintended side effect, but please
correct me if you have better understanding of the code and its history.