Changing clang reflection labels (#8044)
There are two clang reflection implementations:

* The one from Matúš that implements the Reflection TS
* The one from Dan that implements p2996

The latter is what will be in C++26, but that's hard to tell from the
labels since the former is labeled "reflection" while the latter is
labeled "experimental p2996". This changes the labels to "reflection -
TS" and "relfection - C++26", respectively, to make it clearer as to
which is which.