Send a "request compile" if compile on demand isn't set (#8065)
This fixes #8064 - hitting ctrl-enter worked _if compile on demand was
set_. Otherwise the compiler ignored the editor change.

We still need to emit the change (else the compiler window and tools may
not notice any changed text), but then if compile on demand is not on,
we manually trigger a request.