Improved diff view information
With this patch diff view honours the pane name when selecting lhs and rhs.
Now the option name is the pane name rather than the compiler. The compiler
name and options are now displayed below the pane name in the dropdown.
If a pane name is not specified everything defaults as before.