Ensure documentation for LLVM IR casting operations can be found
Previously, if the LLVM IR Viewer contains a casting operation in the
code, the 'View IR Documentation' functionality would be unable to
find the operation's docs. This is because the generated documentation
adds the '-to' suffix to all casting operations, causing a mismatch
with the instruction's real name in the IR.

The fix is to remove the trailing '-to' from any instruction parsed
from the LLVM IR docs, as it is probably a casting operation.

We'll also add ptrtoaddr to the llvmStatement list, as it didn't
seem to be recognized as an operation otherwise (and so failed
the precondition to have a 'View IR Documentation link).