perl: work around a bug (or bugs) in the B module
The B module is the deep internals interface for Perl values and code,
and B::Concise uses that to extract the information about each OP that
it dumps.

Unfortunately the B::COP label() method isn't properly marking labels
stored as Unicode as such when it returns them.

Work around that by re-encoding them.

https://github.com/Perl/perl5/issues/24040