perl: add perl-concise mode for formatting B::Concise output
This mostly just ensures the opcode is marked correctly for asm-docs,
and ensures that the UI recognizes them afterwards (which took some
digging.)

The formatting just makes sure that opcodes are properly highlighted,
which was hit and miss with asm-mode, otherwise it is pretty bad.  I
spent a fair amount of time trying to improve it but didn't make any
significant improvements.